Move all from advxml.core.validate to advxml.core
