+++ src/main/scala/advxml/core/MonadEx.scala
@@ -1,0 +1,9 @@
+package advxml.core
+
+object MonadEx {
+  def apply[F[_]: MonadEx]: MonadEx[F] = implicitly[MonadEx[F]]
+}
+
+object MonadNelEx {
+  def apply[F[_]: MonadNelEx]: MonadNelEx[F] = implicitly[MonadNelEx[F]]
+}
